Added Niharika's calibration tools with "calibrate.py"
Added Niharika's spectrum calibration tools/functions and descriptions to "calibration.py".
